backend/requirements.txt
requests
beautifulsoup4
langchain
langchain-text-splitters
qdrant-client
sentence-transformers
fastapi
uvicorn
python-dotenv
google-generativeai
python-jose[cryptography]
python-multipart 